中間発表後

+++ destroy.php
@@ -2,11 +2,15 @@
 
     require_once 'bootstrap.php';
 
-    if(isset($_REQUEST['id'])) {
-          $id = $_REQUEST['id'];
-          $article = Article::find($id);
-          $article->delete();
+    if (!($_SESSION['username'])){
+      header('Location: index.php');
+      exit;
     }
 
+    if(isset($_REQUEST['id'])) {
+      $id = $_REQUEST['id'];
+      $article = Article::find($id);
+      $article->delete();
+}
     header('Location:index.php');
     exit;
